chore(api): Move lifetime_usages_controller.rb to subfolder (#3814)
The lifetime_usage resource is nested under subscription, just like
alerting or upcoming entitlements.
Nested controller should be under the nested folder.

Publicly, this does not change the routes at all.
